ProblematicPublicKeys - Problem 3
Problem 4 from
http://eagle.cs.mercer.edu/MUSpringContest/MUSpring2014.pdf

The Problem
On February 15, 2012, the New York Times reported a flaw in the method
of generating keys for
a public­key encryption system (“Researchers Find a Flaw in a Widely
Used Online Encryption
Method” by John Markoff). A public key is essentially the product of
two unique prime numbers.
Unfortunately, the prime numbers being used for generating keys were
not always unique. This
flaw enables an attacker to determine private keys given a set of
flawed public keys simply by
finding the shared factors.
Your job is to write a program that takes flawed public keys and
determines the corresponding
private keys. For the purposes of this problem, a private key consists
of a pair of prime numbers,
2 < K1
, K2 < 2
31
, and the corresponding public key consists of their product K1
* K2
.
Input
The first line of input consists of a single integer P, (1 ≤ P ≤ 100),
which is the number of data
sets that follow. Each data set consists of two or more lines. The
first line of the data set
contains two integer values, N (1 ≤ N ≤ P) and M (2 ≤ M ≤ 100),
separated by a space. N is the
data set number, and M is the number of data values that follow. The
following ⌈M/5⌉ lines
contain M “public keys,” 5 keys per line (except possibly the last),
each key, ki
(0 ≤ i < M,
5 < ki < 2
31
), the product of exactly 2 prime numbers.
Output
For each data set there are several lines of output. The first line
contains only the data set
number. The following lines contain the prime factors of the input data
values in ascending
numeric order, 5 per line, except for the last line. Output values on
the same line are separated
by a space.

Sample Input
3
1 6
221 391 713 1457 901
299
2 4
13193 18721 31897 18527
3 2
2143650557 2140117121
Sample Output
1
13 17 23 31 47
53
2
79 97 167 191 193
3
32717 65413 65521