Problem 5 (SubDiagnol Paths)
Sub­Diagonal Paths (prob2 from mercer 2014)
The Problem
You are to find all of the paths on the bottom diagonal of a n x n
grid. The path must only go
from left­to­right or bottom­to­top. Given a dimension of the grid (ex.
n = 4), specify the number
of such paths (ex. solution = 14). (Image below is from Wikipedia):
Input
The input consists of a single integer n, the dimension of the square
grid, 1 <= n <= 30, on each
line. A zero will indicate the end of the input and should not be
processed.
Output
For each input n, you should output the number of paths, as described
above, that exist in an nxn
grid, one per line.
Sample Input
1
2
3
4
0
Sample Output
1
2
5
14